Brewery/Tag/Review Controllers, DAOs, and IDAOs work. Endpoints all work.

+++ dotnet/Capstone/Models/Review.cs
@@ -9,7 +9,7 @@ namespace Capstone.Models
     {
         public int ReviewId { get; set; }
         public int BreweryId { get; set; }
-        public int Rating { get; set; }
+        public decimal Rating { get; set; }
         public string Content { get; set; }
     }
 }
